src/main/java/models/Student.java
package models;

public class Student {
    private String name;
    private String email;
    private String phoneNumber;
    private int courseId;
    private String enrollmentDate;

    private int id;


    public Student(String name, String email, String phoneNumber, int courseId, String enrollmentDate ){
        this.name = name;
        this.email = email;
        this.phoneNumber = phoneNumber;
        this.courseId = courseId;
        this.enrollmentDate = enrollmentDate;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getEmail() {
        return email;
    }

    public void setEmail(String email) {
        this.email = email;
    }

    public String getPhoneNumber() {
        return phoneNumber;
    }

    public void setPhoneNumber(String phoneNumber) {
        this.phoneNumber = phoneNumber;
    }

    public int getCourseId() {
        return courseId;
    }

    public void setCourseId(int courseId) {
        this.courseId = courseId;
    }

    public String getEnrollmentDate() {
        return enrollmentDate;
    }

    public void setEnrollmentDate(String enrollmentDate) {
        this.enrollmentDate = enrollmentDate;
    }

    public int getId() {
        return id;
    }

    public void setId(int id) {
        this.id = id;
    }
}
